Tested on 4.1.0, streamDemo added

+++ unabto_harmony_platform/unabto_harmony_crypto.c
@@ -66,17 +66,17 @@ void unabto_hmac_sha256_buffers(const unabto_buffer keys[], uint8_t keysSize,
 
  
     for (i = 0; i < messagesSize; i++) {
-        uint8_t size = messages[i].size;
+        uint16_t size = messages[i].size;
         uint8_t* data = messages[i].data;
         if (size > BUFFER_SIZE) {
             return;
         }
-        // workaround for bug, if the data is not copied over to the buffer the algorithm fails.
         memcpy(buffer, data, size);
         status = CRYPT_HMAC_DataAdd(&ctx, buffer, size);
         if (status != 0) {
             return;
         }
+        // workaround for bug, if the data is not copied over to the buffer the algorithm fails.
     }
     
     uint8_t hmac[CRYPT_SHA256_DIGEST_SIZE];
